+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/symbols/internal/asm/ClassStubTest.java
@@ -21,10 +21,10 @@ import java.util.List;
 
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.jetbrains.annotations.NotNull;
+import org.junit.jupiter.api.AutoClose;
 import org.junit.jupiter.api.Test;
 import org.pcollections.PSet;
 
-import net.sourceforge.pmd.lang.java.JavaParsingHelper;
 import net.sourceforge.pmd.lang.java.symbols.JClassSymbol;
 import net.sourceforge.pmd.lang.java.symbols.JConstructorSymbol;
 import net.sourceforge.pmd.lang.java.symbols.JRecordComponentSymbol;
@@ -35,15 +35,25 @@ import net.sourceforge.pmd.lang.java.types.JTypeMirror;
 import net.sourceforge.pmd.lang.java.types.Substitution;
 import net.sourceforge.pmd.lang.java.types.TypeSystem;
 import net.sourceforge.pmd.util.CollectionUtil;
+import net.sourceforge.pmd.util.PmdClasspathConfig;
 
 class ClassStubTest {
+
+    @AutoClose
+    private PmdClasspathConfig.OpenClasspath classpath =
+        PmdClasspathConfig.pmdClasspath().open();
+
+    private @NotNull TypeSystem newTypeSystem() {
+        return TypeSystem.usingClasspath(Classpath.forOpenClasspath(classpath));
+    }
+
     // while parsing the annotation type, ClassStub's parseLock.ensureParsed()
     // is called multiple times, reentering the parselock while the status is
     // still BEING_PARSED.
     @Test
     void loadAndParseAnnotation() {
         // class stub - annotation type
-        TypeSystem ts = TypeSystem.usingClassLoaderClasspath(JavaParsingHelper.class.getClassLoader());
+        TypeSystem ts = newTypeSystem();
         JClassSymbol classSymbol = ts.getClassSymbol("java.lang.Deprecated");
         PSet<String> annotationAttributeNames = classSymbol.getAnnotationAttributeNames();
         assertFalse(annotationAttributeNames.isEmpty());
@@ -52,7 +62,7 @@ class ClassStubTest {
 
     @Test
     void recordReflectionTest() {
-        TypeSystem ts = TypeSystem.usingClassLoaderClasspath(JavaParsingHelper.class.getClassLoader());
+        TypeSystem ts = newTypeSystem();
         JClassSymbol pointRecord = loadRecordClass(ts, "Point");
         List<JRecordComponentSymbol> components = pointRecord.getRecordComponents();
         assertThat(components, hasSize(2));
@@ -71,7 +81,7 @@ class ClassStubTest {
 
     @Test
     void varargsRecordReflectionTest() {
-        TypeSystem ts = TypeSystem.usingClassLoaderClasspath(JavaParsingHelper.class.getClassLoader());
+        TypeSystem ts = newTypeSystem();
         JClassSymbol record = loadRecordClass(ts, "Varargs");
         List<JRecordComponentSymbol> components = record.getRecordComponents();
         assertThat(components, hasSize(1));
@@ -88,7 +98,7 @@ class ClassStubTest {
 
     @Test
     void annotatedRecordReflectionTest() {
-        TypeSystem ts = TypeSystem.usingClassLoaderClasspath(JavaParsingHelper.class.getClassLoader());
+        TypeSystem ts = newTypeSystem();
         JClassSymbol record = loadRecordClass(ts, "Annotated");
         List<JRecordComponentSymbol> components = record.getRecordComponents();
         assertThat(components, hasSize(2));
@@ -114,7 +124,7 @@ class ClassStubTest {
 
     @Test
     void targetRecordComponentReflectionTest() {
-        TypeSystem ts = TypeSystem.usingClassLoaderClasspath(JavaParsingHelper.class.getClassLoader());
+        TypeSystem ts = newTypeSystem();
         JClassSymbol record = loadRecordClass(ts, "AnnotatedForRecord");
         List<JRecordComponentSymbol> components = record.getRecordComponents();
         assertThat(components, hasSize(1));
@@ -126,7 +136,7 @@ class ClassStubTest {
 
     @Test
     void testLoadScalaPrivateClassInInterface() {
-        TypeSystem ts = TypeSystem.usingClassLoaderClasspath(JavaParsingHelper.class.getClassLoader());
+        TypeSystem ts = newTypeSystem();
         JClassSymbol itf = loadScalaClass(ts, "InterfaceWithPrivateInner");
         assertThat(itf, hasProperty("interface", equalTo(true)));
         assertThat(itf.getDeclaredClasses(), hasSize(2));
@@ -143,7 +153,7 @@ class ClassStubTest {
 
     @Test
     void testLoadAnonClassFromEnum() {
-        TypeSystem ts = TypeSystem.usingClassLoaderClasspath(JavaParsingHelper.class.getClassLoader());
+        TypeSystem ts = newTypeSystem();
         JClassSymbol enumClass = loadTestDataClass(ts, "EnumConstantWithBody");
         assertThat(enumClass, hasProperty("simpleName", equalTo("EnumConstantWithBody")));
 
@@ -157,7 +167,7 @@ class ClassStubTest {
 
     @Test
     void testLoadLocalClass() {
-        TypeSystem ts = TypeSystem.usingClassLoaderClasspath(JavaParsingHelper.class.getClassLoader());
+        TypeSystem ts = newTypeSystem();
         JClassSymbol outerClass = loadTestDataClass(ts, "LocalClasses");
 
         AsmSymbolResolver resolver = (AsmSymbolResolver) ts.bootstrapResolver();